pyproject.toml
[tool.poetry]
name = "place_info_bot"
version = "0.1.0"
description = ""
authors = ["Spartan327 <vbutovchenko@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = {extras = ["socks"], version = "^13.7"}
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
